rename file to lowercase
Comment: https://github.com/htmlacademy-ecmascript/461957-taskmanager-9/pull/5#discussion_r317359014
